+++ xcube/webapi/context.py
@@ -24,7 +24,8 @@ import logging
 import os
 import os.path
 import threading
-from typing import Any, Dict, List, Optional, Tuple, Callable, Collection, Sequence
+from typing import Any, Dict, List, Optional, Tuple, Callable, Collection, Set
+from typing import Sequence
 
 import fiona
 import numpy as np
@@ -33,18 +34,29 @@ import s3fs
 import xarray as xr
 import zarr
 
-from xcube.constants import FORMAT_NAME_ZARR, FORMAT_NAME_NETCDF4, FORMAT_NAME_LEVELS
+from xcube.constants import FORMAT_NAME_LEVELS
+from xcube.constants import FORMAT_NAME_NETCDF4
+from xcube.constants import FORMAT_NAME_ZARR
 from xcube.core.dsio import guess_dataset_format
-from xcube.core.mldataset import FileStorageMultiLevelDataset, BaseMultiLevelDataset, MultiLevelDataset, \
-    ComputedMultiLevelDataset, ObjectStorageMultiLevelDataset
+from xcube.core.mldataset import BaseMultiLevelDataset
+from xcube.core.mldataset import ComputedMultiLevelDataset
+from xcube.core.mldataset import FileStorageMultiLevelDataset
+from xcube.core.mldataset import MultiLevelDataset
+from xcube.core.mldataset import ObjectStorageMultiLevelDataset
 from xcube.core.verify import assert_cube
 from xcube.util.cache import MemoryCacheStore, Cache
 from xcube.util.cmaps import get_cmap
 from xcube.util.perf import measure_time
 from xcube.util.tilegrid import TileGrid
 from xcube.version import version
-from xcube.webapi.defaults import DEFAULT_CMAP_CBAR, DEFAULT_CMAP_VMIN, DEFAULT_CMAP_VMAX, DEFAULT_TRACE_PERF
-from xcube.webapi.errors import ServiceConfigError, ServiceError, ServiceBadRequestError, ServiceResourceNotFoundError
+from xcube.webapi.defaults import DEFAULT_CMAP_CBAR
+from xcube.webapi.defaults import DEFAULT_CMAP_VMAX
+from xcube.webapi.defaults import DEFAULT_CMAP_VMIN
+from xcube.webapi.defaults import DEFAULT_TRACE_PERF
+from xcube.webapi.errors import ServiceBadRequestError
+from xcube.webapi.errors import ServiceConfigError
+from xcube.webapi.errors import ServiceError
+from xcube.webapi.errors import ServiceResourceNotFoundError
 from xcube.webapi.reqparams import RequestParams
 
 COMPUTE_DATASET = 'compute_dataset'
@@ -145,6 +157,42 @@ class ServiceContext:
     @property
     def trace_perf(self) -> bool:
         return self._trace_perf
+
+    @property
+    def access_control(self) -> Dict[str, Any]:
+        return dict(self._config.get('AccessControl', {}))
+
+    @property
+    def required_scopes(self) -> List[str]:
+        return self.access_control.get('RequiredScopes', [])
+
+    def get_required_dataset_scopes(self,
+                                    dataset_descriptor: DatasetDescriptor) -> Set[str]:
+        return self._get_required_scopes(dataset_descriptor, 'read:dataset', 'Dataset',
+                                         dataset_descriptor['Identifier'])
+
+    def get_required_variable_scopes(self,
+                                     dataset_descriptor: DatasetDescriptor,
+                                     var_name: str) -> Set[str]:
+        return self._get_required_scopes(dataset_descriptor, 'read:variable', 'Variable',
+                                         var_name)
+
+    def _get_required_scopes(self,
+                             dataset_descriptor: DatasetDescriptor,
+                             base_scope: str,
+                             value_name: str,
+                             value: str) -> Set[str]:
+        base_scope_prefix = base_scope + ':'
+        pattern_scope = base_scope_prefix + '{' + value_name + '}'
+        dataset_access_control = dataset_descriptor.get('AccessControl', {})
+        dataset_required_scopes = dataset_access_control.get('RequiredScopes', [])
+        dataset_required_scopes = set(self.required_scopes + dataset_required_scopes)
+        dataset_required_scopes = {scope for scope in dataset_required_scopes
+                                   if scope == base_scope or scope.startswith(base_scope_prefix)}
+        if pattern_scope in dataset_required_scopes:
+            dataset_required_scopes.remove(pattern_scope)
+            dataset_required_scopes.add(base_scope_prefix + value)
+        return dataset_required_scopes
 
     def get_service_url(self, base_url, *path: str):
         if self._prefix:
@@ -341,14 +389,14 @@ class ServiceContext:
         place_group_id = place_group_descriptor.get("PlaceGroupRef")
         if place_group_id:
             if is_global:
-                raise ServiceError("'PlaceGroupRef' cannot be used in a global place group")
+                raise ServiceConfigError("'PlaceGroupRef' cannot be used in a global place group")
             if len(place_group_descriptor) > 1:
-                raise ServiceError("'PlaceGroupRef' if present, must be the only entry in a 'PlaceGroups' item")
+                raise ServiceConfigError("'PlaceGroupRef' if present, must be the only entry in a 'PlaceGroups' item")
             return self.get_global_place_group(place_group_id, base_url, load_features=load_features)
 
         place_group_id = place_group_descriptor.get("Identifier")
         if not place_group_id:
-            raise ServiceError("Missing 'Identifier' entry in a 'PlaceGroups' item")
+            raise ServiceConfigError("Missing 'Identifier' entry in a 'PlaceGroups' item")
 
         if place_group_id in self._place_group_cache:
             place_group = self._place_group_cache[place_group_id]
@@ -357,7 +405,7 @@ class ServiceContext:
 
             place_path_wc = place_group_descriptor.get("Path")
             if not place_path_wc:
-                raise ServiceError("Missing 'Path' entry in a 'PlaceGroups' item")
+                raise ServiceConfigError("Missing 'Path' entry in a 'PlaceGroups' item")
             if not os.path.isabs(place_path_wc):
                 place_path_wc = os.path.join(self._base_dir, place_path_wc)
             source_paths = glob.glob(place_path_wc)
@@ -395,7 +443,7 @@ class ServiceContext:
 
             sub_place_group_configs = place_group_descriptor.get("Places")
             if sub_place_group_configs:
-                raise ServiceError("Invalid 'Places' entry in a 'PlaceGroups' item: not implemented yet")
+                raise ServiceConfigError("Invalid 'Places' entry in a 'PlaceGroups' item: not implemented yet")
             # sub_place_group_descriptors = place_group_config.get("Places")
             # if sub_place_group_descriptors:
             #     sub_place_groups = self._load_place_groups(sub_place_group_descriptors)
@@ -454,12 +502,9 @@ class ServiceContext:
                 feature_index[property_value] = feature
         return feature_index
 
-
     @classmethod
     def _remove_feature_id(cls, feature: Dict):
         cls._remove_id(feature)
-        # if "properties" in feature:
-        #    cls._remove_id(feature["properties"])
 
     @classmethod
     def _remove_id(cls, properties: Dict):